+++ modules/client/src/internal/schemas.ts
@@ -169,3 +169,21 @@ export const DaoUpdateSchema = object({
   initData: Uint8ArraySchema.notRequired(),
   daoFactoryAddress: AddressOrEnsSchema.notRequired(),
 });
+
+export const IsPluginUpdateValidSchema = object({
+  actions: array().of(object({
+    to: AddressOrEnsSchema.required(),
+    value: BigintSchema.required(),
+    data: Uint8ArraySchema.required(),
+  })).required().min(1),
+  daoAddress: AddressOrEnsSchema.required(),
+});
+export const IsDaoUpdateValidSchema = object({
+  actions: array().of(object({
+    to: AddressOrEnsSchema.required(),
+    value: BigintSchema.required(),
+    data: Uint8ArraySchema.required(),
+  })).required().min(1),
+  daoAddress: AddressOrEnsSchema.required(),
+  version: array().of(number()).length(3).notRequired(),
+});